Split of Country and Region
