+++ src/countries
@@ -1,13 +1,13 @@
-p({"class":"foam.nanos.auth.Country", "code":"BR", "ISOCode":"BRA","name":"Brazil"})
-p({"class":"foam.nanos.auth.Country", "code":"CA", "ISOCode":"CAN","name":"Canada"})
-p({"class":"foam.nanos.auth.Country", "code":"CN", "ISOCode":"CHN","name":"China", "alternativeName":["People's Republic of China"]})
-p({"class":"foam.nanos.auth.Country", "code":"IN", "ISOCode":"IND","name":"India"})
-p({"class":"foam.nanos.auth.Country", "code":"JM", "ISOCode":"JAM","name":"Jamaica"})
-p({"class":"foam.nanos.auth.Country", "code":"LB", "ISOCode":"LSO","name":"Lebanon"})
-p({"class":"foam.nanos.auth.Country", "code":"MX", "ISOCode":"MEX","name":"Mexico"})
-p({"class":"foam.nanos.auth.Country", "code":"MY", "ISOCode":"MYS","name":"Malaysia"})
-p({"class":"foam.nanos.auth.Country", "code":"RS", "ISOCode":"SRB","name":"Serbia"})
-p({"class":"foam.nanos.auth.Country", "code":"TT", "ISOCode":"TTO","name":"Trinidad and Tobago"})
-p({"class":"foam.nanos.auth.Country", "code":"GB", "ISOCode":"GBR","name":"United Kingdom", "alternativeName":["Great Britain"]})
-p({"class":"foam.nanos.auth.Country", "code":"US", "ISOCode":"USA","name":"USA", "alternativeName":["United States", "United States of America"]})
-p({"class":"foam.nanos.auth.Country", "code":"ZA", "ISOCode":"ZAF","name":"South Africa"})
+p({"class":"foam.nanos.auth.Country", "code":"BR", "ISO_3166_1":"BRA","name":"Brazil"})
+p({"class":"foam.nanos.auth.Country", "code":"CA", "ISO_3166_1":"CAN","name":"Canada"})
+p({"class":"foam.nanos.auth.Country", "code":"CN", "ISO_3166_1":"CHN","name":"China", "alternativeNames":["People's Republic of China"]})
+p({"class":"foam.nanos.auth.Country", "code":"IN", "ISO_3166_1":"IND","name":"India"})
+p({"class":"foam.nanos.auth.Country", "code":"JM", "ISO_3166_1":"JAM","name":"Jamaica"})
+p({"class":"foam.nanos.auth.Country", "code":"LB", "ISO_3166_1":"LSO","name":"Lebanon"})
+p({"class":"foam.nanos.auth.Country", "code":"MX", "ISO_3166_1":"MEX","name":"Mexico"})
+p({"class":"foam.nanos.auth.Country", "code":"MY", "ISO_3166_1":"MYS","name":"Malaysia"})
+p({"class":"foam.nanos.auth.Country", "code":"RS", "ISO_3166_1":"SRB","name":"Serbia"})
+p({"class":"foam.nanos.auth.Country", "code":"TT", "ISO_3166_1":"TTO","name":"Trinidad and Tobago"})
+p({"class":"foam.nanos.auth.Country", "code":"GB", "ISO_3166_1":"GBR","name":"United Kingdom", "alternativeName":["Great Britain"]})
+p({"class":"foam.nanos.auth.Country", "code":"US", "ISO_3166_1":"USA","name":"USA", "alternativeNames":["United States", "United States of America"]})
+p({"class":"foam.nanos.auth.Country", "code":"ZA", "ISO_3166_1":"ZAF","name":"South Africa"})
