+++ src/services
@@ -35,13 +35,14 @@ p({"class":"foam.nanos.boot.NSpec", "name":"login",                            "
 p({"class":"foam.nanos.boot.NSpec", "name":"memory",                           "lazy":true,  "serve":false, "serviceScript":"return new foam.nanos.http.RefreshWebAgent(new foam.nanos.http.MemoryWebAgent());"})
 //p({"class":"foam.nanos.boot.NSpec", "name":"countryDAO",                       "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.EasyDAO.Builder(x).setPm(true).setAuthenticateRead(false).setJournaled(true).setJournalName(\"countries\").setOf(foam.nanos.auth.Country.getOwnClassInfo()).build().orderBy(foam.nanos.auth.Country.NAME);","client":"{\"of\":\"foam.nanos.auth.Country\"}"})
 p({"class":"foam.nanos.boot.NSpec", "name":"countryDAO",                       "lazy":true,  "serve":true,  "authenticate": false, "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.AuthenticatedDAO(\"countryDAO\", false, new foam.dao.java.JDAO(x, foam.nanos.auth.Country.getOwnClassInfo(), \"countries\"))).orderBy(foam.nanos.auth.Country.NAME);","client":"{\"of\":\"foam.nanos.auth.Country\"}"})
-p({"class":"foam.nanos.boot.NSpec", "name":"countryAndRegionService",           "lazy":true,  "serve":true, "serviceClass":"foam.nanos.auth.CountryAndRegionCacheService"})
+p({"class":"foam.nanos.boot.NSpec", "name":"countryService",                   "lazy":true,  "serve":true, "serviceClass":"foam.nanos.auth.CountryService"})
 
 //p({"class":"foam.nanos.boot.NSpec", "name":"languageDAO",                      "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.EasyDAO.Builder(x).setPm(true).setAuthenticateRead(false).setJournaled(true).setJournalName(\"languages\").setOf(foam.nanos.auth.Language.getOwnClassInfo()).build();", "client":"{\"of\":\"foam.nanos.auth.Language\"}"})
 p({"class":"foam.nanos.boot.NSpec", "name":"languageDAO",                      "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.java.JDAO(x, foam.nanos.auth.Language.getOwnClassInfo(), \"languages\"));","client":"{\"of\": \"foam.nanos.auth.Language\"}"})
 
 //p({"class":"foam.nanos.boot.NSpec", "name":"regionDAO",                        "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.EasyDAO.Builder(x).setPm(true).setAuthenticateRead(false).setJournaled(true).setJournalName(\"regions\").setOf(foam.nanos.auth.Region.getOwnClassInfo()).build().orderBy(foam.nanos.auth.Region.NAME);","client":"{\"of\":\"foam.nanos.auth.Region\"}"})
 p({"class":"foam.nanos.boot.NSpec", "name":"regionDAO",                        "lazy":true,  "serve":true,  "authenticate": false, "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.AuthenticatedDAO(\"regionDAO\", false, new foam.dao.java.JDAO(x, foam.nanos.auth.Region.getOwnClassInfo(), \"regions\"))).orderBy(foam.nanos.auth.Region.NAME);","client":"{\"of\":\"foam.nanos.auth.Region\"}"})
+p({"class":"foam.nanos.boot.NSpec", "name":"regionService",                    "lazy":true,  "serve":true, "serviceClass":"foam.nanos.auth.RegionService"})
 
 //p({"class":"foam.nanos.boot.NSpec", "name":"exportDriverRegistryDAO",          "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.EasyDAO.Builder(x).setPm(true).setJournaled(true).setJournalName(\"exportDriverRegistrys\").setOf(foam.nanos.export.ExportDriverRegistry.getOwnClassInfo()).build();", "client":"{\"of\":\"foam.nanos.export.ExportDriverRegistry\"}"})
 p({"class":"foam.nanos.boot.NSpec", "name":"exportDriverRegistryDAO",          "lazy":true,  "serve":true,  "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.java.JDAO(x, foam.nanos.export.ExportDriverRegistry.getOwnClassInfo(), \"exportDriverRegistrys\"));","client":"{\"of\":\"foam.nanos.export.ExportDriverRegistry\"}"})
